refactor(samples): move greetingGet to a separate secure schema
