+++ samples/chaotic_openapi_service/handlers/insecure/openapi.yaml
@@ -28,17 +28,6 @@ paths:
                         application/json:
                             schema:
                                 $ref: '#/components/schemas/GreetingResponse'
-
-    /secure/greeting:
-        get:
-            operationId: greetingGet
-            responses:
-                '200':
-                    description: OK
-                    content:
-                        application/json:
-                            schema:
-                                $ref: '#/components/schemas/GreetingResponse'
 components:
     schemas:
         GreetingResponse:
