+++ samples/chaotic_openapi_service/handlers/secure/openapi.yaml
@@ -1,0 +1,25 @@
+openapi: 3.0.0
+info:
+    title: Secure handler
+    version: '1.0'
+paths:
+    /secure/greeting:
+        get:
+            operationId: greetingGet
+            responses:
+                '200':
+                    description: OK
+                    content:
+                        application/json:
+                            schema:
+                                $ref: '#/components/schemas/GreetingResponse'
+components:
+    schemas:
+        GreetingResponse:
+            type: object
+            additionalProperties: false
+            required:
+              - greeting
+            properties:
+                greeting:
+                    type: string
