+++ samples/chaotic_openapi_service/src/ya.make
@@ -19,7 +19,7 @@ SRCS(
     hello_handler.cpp
     say_hello.cpp
     handlers/insecure/insecuresecretpost/view.cpp
-    handlers/insecure/greetingget/view.cpp
+    handlers/secure/greetingget/view.cpp
 )
 
 ADDINCL(
@@ -72,7 +72,6 @@ RUN_PROGRAM(
         ${CHAOTIC_INCLUDES}
         ${CHAOTIC_OPENAPI_INCLUDES}
         handlers/insecure/insecuresecretpost/view.hpp
-        handlers/insecure/greetingget/view.hpp
     IN_NOPARSE
         ../handlers/insecure/openapi.yaml
     OUT
@@ -92,13 +91,38 @@ RUN_PROGRAM(
         src/handlers/insecure/insecuresecretpost/requests.cpp
         src/handlers/insecure/insecuresecretpost/responses.cpp
 
-        include/handlers/insecure/greetingget/handler.hpp
-        include/handlers/insecure/greetingget/requests.hpp
-        include/handlers/insecure/greetingget/responses.hpp
+        config.chaotic.yaml
+)
 
-        src/handlers/insecure/greetingget/handler.cpp
-        src/handlers/insecure/greetingget/requests.cpp
-        src/handlers/insecure/greetingget/responses.cpp
+RUN_PROGRAM(
+    taxi/uservices/userver/chaotic-openapi/bin
+        --name secure
+        --gen handlers
+        -o ${BINDIR}/handlers/secure
+        --clang-format ''
+        ../handlers/secure/openapi.yaml
+    OUTPUT_INCLUDES
+        ${CHAOTIC_INCLUDES}
+        ${CHAOTIC_OPENAPI_INCLUDES}
+        handlers/secure/greetingget/view.hpp
+    IN_NOPARSE
+        ../handlers/secure/openapi.yaml
+    OUT
+        include/handlers/secure/openapi.hpp
+        include/handlers/secure/openapi_fwd.hpp
+        include/handlers/secure/openapi_parsers.ipp
+        include/handlers/secure/openapi_sax_parsers.hpp
+
+        src/handlers/secure/openapi.cpp
+
+        include/handlers/secure/greetingget/handler.hpp
+        include/handlers/secure/chaotic_handlers_list.hpp
+        include/handlers/secure/greetingget/requests.hpp
+        include/handlers/secure/greetingget/responses.hpp
+
+        src/handlers/secure/greetingget/handler.cpp
+        src/handlers/secure/greetingget/requests.cpp
+        src/handlers/secure/greetingget/responses.cpp
 
         config.chaotic.yaml
 )
@@ -106,10 +130,12 @@ RUN_PROGRAM(
 RUN_PROGRAM(
     taxi/uservices/userver/scripts/chaotic
         ${BINDIR}/config.chaotic.yaml
+        ${BINDIR}/handlers/secure/config.chaotic.yaml
         ${CURDIR}/../static_config.yaml
         -o ./config.yaml
     IN_NOPARSE
         ${BINDIR}/config.chaotic.yaml
+        ${BINDIR}/handlers/secure/config.chaotic.yaml
         ../static_config.yaml
     OUT
         config.yaml
